Add order status function documentation to iPay ACBA plugin
A new function to get the status of the ACBA iPay order has been added to the class-ipay-acba-gateway.php file. This addition enhances control and tracking capabilities, allowing developers to check the status of any order via its ID, returning either an object or 'false' if the request is not successful.